chore(types): fix all mypy type errors in Phase 2 core and engines
- Install pandas-stubs and types-PyYAML for better type coverage
- Fix Module/None assignments with type: ignore comments
- Change optimal_size_range from tuple[int, int] to tuple[float, float]
- Add type ignores for Engine return types (DataFrameLike protocol)
- Fix DataFrame constructor overload issues
- Fix max() key function in heuristics
- Fix AvroReader/AvroWriter optional imports
- Fix missing return statement in reader.py
- Add type ignores for __len__, __getitem__, shape, and columns

All 66 Phase 2 tests passing, mypy clean on core_v2 and engines

+++ src/parquetframe/core_v2/base.py
@@ -16,7 +16,7 @@ try:
 
     POLARS_AVAILABLE = True
 except ImportError:
-    pl = None
+    pl = None  # type: ignore[assignment]
     POLARS_AVAILABLE = False
 
 try:
@@ -24,7 +24,7 @@ try:
 
     DASK_AVAILABLE = True
 except ImportError:
-    dd = None
+    dd = None  # type: ignore[assignment]
     DASK_AVAILABLE = False
 
 
@@ -103,7 +103,7 @@ class EngineCapabilities:
         name: str,
         is_lazy: bool = False,
         supports_distributed: bool = False,
-        optimal_size_range: tuple[int, int] = (0, float("inf")),
+        optimal_size_range: tuple[float, float] = (0, float("inf")),
         memory_efficiency: float = 1.0,
         performance_score: float = 1.0,
     ):
